Fixed stupid OnStopServer stupidly not being called (fix Subscribe/Unsubscribe failure)
